layout/forms/test/test_bug36619.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=36619
-->
<head>
  <title>Test for Bug 36619</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=36619">Mozilla Bug 36619</a>
<p id="display"></p>
<div id="content" style="display:none;">
  <input id='a' type='file'>
</div>
<button id='b' onclick="document.getElementById('a').click();">Show Filepicker</button>
<button id='c' onmousedown="document.getElementById('a').click();">Show Filepicker</button>
<button id='d' onmouseup="document.getElementById('a').click();">Show Filepicker</button>
<pre id="test">
<script type="application/javascript;version=1.8">

/** Test for Bug 36619 **/

let gGen = doTest();
function continueTest() {
  gGen.next();
}

function waitForCondition(condition, errorMsg) {
  var tries = 0;
  var interval = setInterval(function() {
    if (tries >= 10) {
      ok(false, errorMsg);
      moveOn();
    }
    if (condition()) {
      moveOn();
    }
    tries++;
  }, 100);
  var moveOn = function() {
    clearInterval(interval);
    try {
      continueTest();
    } catch (e if e instanceof StopIteration) {
      SimpleTest.finish();
    }
  };
}

SimpleTest.waitForExplicitFinish();

var MockFilePicker = SpecialPowers.MockFilePicker;
MockFilePicker.init(window);

// enable popups the first time
SpecialPowers.pushPrefEnv({'set': [
  ["dom.disable_open_during_load", true],
  ["privacy.popups.showBrowserMessage", false]
]}, function() {
  SimpleTest.waitForFocus(function() {
    continueTest();
  });
});

let showing = function() MockFilePicker.showing;
let notShowing = function() !MockFilePicker.showing;

function doTest() {
  ok(true, "Test start");
  var b = document.getElementById('b');
  b.focus(); // Be sure the element is visible.

  synthesizeMouseAtCenter(b, {});
  yield waitForCondition(showing,
                         "File picker showAsync method should have been called");

  MockFilePicker.reset();

  // Tests that a click on 'a' doesn't call the show method, because it is hidden.
  document.getElementById("a").click();
  yield waitForCondition(notShowing,
                         "File picker showAsync method should not have been called");

  MockFilePicker.reset();

  synthesizeMouseAtCenter(document.getElementById('c'), {});
  yield waitForCondition(notShowing,
                         "File picker showAsync method should not have been called");

  MockFilePicker.reset();
  synthesizeMouseAtCenter(document.getElementById('d'), {});

  yield waitForCondition(showing,
                         "File picker showAsync method should have been called");

  ok(true, "Test completed");
  MockFilePicker.cleanup();
}

</script>
</pre>
</body>
</html>
